GBA e-Reader: Split out into separate file
