+++ include/mgba/internal/gba/hardware.h
@@ -16,6 +16,9 @@ CXX_GUARD_START
 
 mLOG_DECLARE_CATEGORY(GBA_HW);
 
+#define EREADER_DOTCODE_STRIDE 1200
+#define EREADER_DOTCODE_SIZE (EREADER_DOTCODE_STRIDE * 40 + 200)
+
 #define IS_GPIO_REGISTER(reg) ((reg) == GPIO_REG_DATA || (reg) == GPIO_REG_DIRECTION || (reg) == GPIO_REG_CONTROL)
 
 struct GBARTCGenericSource {
